2015/02_i_was_told_there_would_be_no_math/main_test.go
package main

import (
	"testing"

	"github.com/SuperDulli/advent-of-code/util"
)

var example = `2x3x4
1x1x10`

func Test_part1(t *testing.T) {
	tests := []struct {
		name  string
		input []string
		want  int
	}{
		{"example", util.SplitLines(example), 58 + 43},
		{"actual", util.ReadLines("input.txt"), 1598415},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part1(tt.input); got != tt.want {
				t.Errorf("part1() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_part2(t *testing.T) {
	tests := []struct {
		name  string
		input []string
		want  int
	}{
		{"example", util.SplitLines(example), 34 + 14},
		{"actual", util.ReadLines("input.txt"), 3812909},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := part2(tt.input); got != tt.want {
				t.Errorf("part2() = %v, want %v", got, tt.want)
			}
		})
	}
}
